requirements.txt
streamlit
pandas
sentence-transformers
faiss-cpu
pypdf
requests
plotly
numpy
scikit-learn